moved renamed container.py to vllm_container.py
